requirements.txt
requests>=2.28.0
pillow>=9.0.0  # opcional, só se quiser ver imagens
